Replaced bridge example test with wiretap one.
