+++ pom.xml
@@ -158,6 +158,10 @@
             <version>4.0.1</version>
             <scope>test</scope>
         </dependency>
+<!--        <dependency>-->
+<!--            <groupId>org.springframework.integration</groupId>-->
+<!--            <artifactId>spring-integration-test</artifactId>-->
+<!--        </dependency>-->
 
     </dependencies>
